Update update-strategies.md
fixes #295 in response to #292

A bit unsure if this still applies @NekR or it should be removed

> With this strategy enabled, index.html (or other files without dynamic name) should be placed in main section of the cache, otherwise they won't be revalidated.
For ServiceWorker this means that only new files will be downloaded and missing files deleted from the cache.
For AppCache it's basically the same as previous strategies since AppCache revalidates all the assets. (HTTP status 304 rule still applies).