Update update-strategies.md (#307)
* Fixes thrown exceptions in apps using postMessage

My app uses postMessage for it's own purposes. This fixes an error that occurs when the appcache updates choke on messages from the application.

* Adds info regarding redirects not working

@NekR

In response to #303

* Update update-strategies.md

fixes #295 in response to #292

A bit unsure if this still applies @NekR or it should be removed

> With this strategy enabled, index.html (or other files without dynamic name) should be placed in main section of the cache, otherwise they won't be revalidated.
For ServiceWorker this means that only new files will be downloaded and missing files deleted from the cache.
For AppCache it's basically the same as previous strategies since AppCache revalidates all the assets. (HTTP status 304 rule still applies).

* Update FAQ.md

as asked for in #227

Also related to #228, but does not add any additional `events` doc

* 4.8.4

* update according to PR comments

* new -> new/changed